requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
dnspython==2.4.1
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.1
filelock==3.12.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
lit==16.0.6
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper @ git+https://github.com/openai/whisper.git@b91c907694f96a3fb9da03d4bbdc83fbcd3a40a4
orjson==3.9.2
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
pydub==0.25.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
